requirements.txt

requests
playwright
python-dotenv
